Remove default interface
