+++ src/shared/clients/cloudWatchLogsClient.ts
@@ -4,15 +4,49 @@
  */
 
 import { CloudWatchLogs } from 'aws-sdk'
+import { ext } from '../extensionGlobals'
+import { ClassToInterface } from '../utilities/tsUtils'
 
-export interface CloudWatchLogsClient {
-    readonly regionCode: string
+export type CloudWatchLogsClient = ClassToInterface<DefaultCloudWatchLogsClient>
+export class DefaultCloudWatchLogsClient {
+    public constructor(public readonly regionCode: string) {}
 
-    describeLogGroups(): AsyncIterableIterator<CloudWatchLogs.LogGroup>
+    public async *describeLogGroups(): AsyncIterableIterator<CloudWatchLogs.LogGroup> {
+        const sdkClient = await this.createSdkClient()
+        const request: CloudWatchLogs.DescribeLogGroupsRequest = {}
+        do {
+            const response = await this.invokeDescribeLogGroups(request, sdkClient)
+            if (response.logGroups) {
+                yield* response.logGroups
+            }
+            request.nextToken = response.nextToken
+        } while (request.nextToken)
+    }
 
-    getLogEvents(request: CloudWatchLogs.GetLogEventsRequest): Promise<CloudWatchLogs.GetLogEventsResponse>
-
-    describeLogStreams(
+    public async describeLogStreams(
         request: CloudWatchLogs.DescribeLogStreamsRequest
-    ): Promise<CloudWatchLogs.DescribeLogStreamsResponse>
+    ): Promise<CloudWatchLogs.DescribeLogStreamsResponse> {
+        const sdkClient = await this.createSdkClient()
+
+        return sdkClient.describeLogStreams(request).promise()
+    }
+
+    public async getLogEvents(
+        request: CloudWatchLogs.GetLogEventsRequest
+    ): Promise<CloudWatchLogs.GetLogEventsResponse> {
+        const sdkClient = await this.createSdkClient()
+
+        return sdkClient.getLogEvents(request).promise()
+    }
+
+    protected async invokeDescribeLogGroups(
+        request: CloudWatchLogs.DescribeLogGroupsRequest,
+        sdkClient: CloudWatchLogs
+    ): Promise<CloudWatchLogs.DescribeLogGroupsResponse> {
+        return sdkClient.describeLogGroups(request).promise()
+    }
+
+    protected async createSdkClient(): Promise<CloudWatchLogs> {
+        return await ext.sdkClientBuilder.createAwsService(CloudWatchLogs, undefined, this.regionCode)
+    }
 }
