+++ src/shared/clients/lambdaClient.ts
@@ -5,41 +5,97 @@
 
 import { Lambda } from 'aws-sdk'
 import { _Blob } from 'aws-sdk/clients/lambda'
+import { ext } from '../extensionGlobals'
+import '../utilities/asyncIteratorShim'
+import { getLogger } from '../logger'
+import { ClassToInterface } from '../utilities/tsUtils'
 
-export interface LambdaClient {
-    readonly regionCode: string
+export type LambdaClient = ClassToInterface<DefaultLambdaClient>
+export class DefaultLambdaClient {
+    public constructor(public readonly regionCode: string) {}
 
-    /**
-     * Deletes Lambda function from AWS
-     * @param name Function to delete
-     */
-    deleteFunction(name: string): Promise<void>
+    public async deleteFunction(name: string): Promise<void> {
+        const sdkClient = await this.createSdkClient()
 
-    /**
-     * Invokes a remote Lambda function
-     * @param name Lambda to invoke
-     * @param payload Payload to invoke with
-     */
-    invoke(name: string, payload?: _Blob): Promise<Lambda.InvocationResponse>
+        const response = await sdkClient
+            .deleteFunction({
+                FunctionName: name,
+            })
+            .promise()
 
-    /**
-     * Lists available Lambda functions. Each iteration represents a single listed function
-     *
-     * TODO: Move to a more predictable iterator: iterate a page of results on each call so each call corresponds to a network request
-     */
-    listFunctions(): AsyncIterableIterator<Lambda.FunctionConfiguration>
+        if (response.$response.error) {
+            throw response.$response.error
+        }
+    }
 
-    /**
-     * Runs AWS.Lambda.getFunction, which gets function metadata (including zip location)
-     * @param name Function name
-     */
-    getFunction(name: string): Promise<Lambda.GetFunctionResponse>
+    public async invoke(name: string, payload?: _Blob): Promise<Lambda.InvocationResponse> {
+        const sdkClient = await this.createSdkClient()
 
-    /**
-     * Runs AWS.Lambda.updateFunctionCode
-     * @param name Function name
-     * @param zipFile Buffer containing zip file data
-     *
-     */
-    updateFunctionCode(name: string, zipFile: Buffer): Promise<Lambda.FunctionConfiguration>
+        const response = await sdkClient
+            .invoke({
+                FunctionName: name,
+                LogType: 'Tail',
+                Payload: payload,
+            })
+            .promise()
+
+        return response
+    }
+
+    public async *listFunctions(): AsyncIterableIterator<Lambda.FunctionConfiguration> {
+        const client = await this.createSdkClient()
+
+        const request: Lambda.ListFunctionsRequest = {}
+        do {
+            const response: Lambda.ListFunctionsResponse = await client.listFunctions(request).promise()
+
+            if (response.Functions) {
+                yield* response.Functions
+            }
+
+            request.Marker = response.NextMarker
+        } while (request.Marker)
+    }
+
+    public async getFunction(name: string): Promise<Lambda.GetFunctionResponse> {
+        getLogger().debug(`GetFunction called for function: ${name}`)
+        const client = await this.createSdkClient()
+
+        try {
+            const response = await client.getFunction({ FunctionName: name }).promise()
+            // prune `Code` from logs so we don't reveal a signed link to customer resources.
+            getLogger().debug('GetFunction returned response (code section pruned): %O', {
+                ...response,
+                Code: 'Pruned',
+            })
+            return response
+        } catch (e) {
+            getLogger().error('Failed to get function: %O', e)
+            throw e
+        }
+    }
+
+    public async updateFunctionCode(name: string, zipFile: Buffer): Promise<Lambda.FunctionConfiguration> {
+        getLogger().debug(`updateFunctionCode called for function: ${name}`)
+        const client = await this.createSdkClient()
+
+        try {
+            const response = await client
+                .updateFunctionCode({
+                    FunctionName: name,
+                    Publish: true,
+                    ZipFile: zipFile,
+                })
+                .promise()
+            getLogger().debug('updateFunctionCode returned response: %O', response)
+            return response
+        } catch (e) {
+            getLogger().error('Failed to run updateFunctionCode: %O', e)
+            throw e
+        }
+    }
+
+    private async createSdkClient(): Promise<Lambda> {
+        return await ext.sdkClientBuilder.createAwsService(Lambda, undefined, this.regionCode)
+    }
 }
