+++ src/test/shared/clients/defaultS3Client.test.ts
@@ -10,7 +10,7 @@ import { ManagedUpload } from 'aws-sdk/lib/s3/managed_upload'
 import { FileStreams } from '../../../shared/utilities/streamUtilities'
 import { anyFunction, anything, capture, deepEqual, instance, mock, verify, when } from '../../utilities/mockito'
 import * as vscode from 'vscode'
-import { DefaultBucket, DefaultFile, DefaultFolder, DefaultS3Client } from '../../../shared/clients/defaultS3Client'
+import { DefaultBucket, DefaultFile, DefaultFolder, DefaultS3Client } from '../../../shared/clients/S3Client'
 import { DEFAULT_DELIMITER, DEFAULT_MAX_KEYS } from '../../../shared/clients/s3Client'
 import { FakeFileStreams } from './fakeFileStreams'
 
@@ -376,7 +376,7 @@ describe('DefaultS3Client', function () {
             const mockResponse: Request<any, AWSError> = mock()
             when(mockS3.listBuckets()).thenReturn(success({ Buckets: [{ Name: bucketName }] }))
             // eslint-disable-next-line @typescript-eslint/unbound-method
-            when(mockResponse.promise).thenReject((undefined as any) as Error)
+            when(mockResponse.promise).thenReject(undefined as any as Error)
             when(mockS3.getBucketLocation(anything())).thenReturn(mockResponse)
 
             const response = await createClient().listBuckets()
@@ -481,12 +481,8 @@ describe('DefaultS3Client', function () {
     })
 
     describe('listObjectVersions', function () {
-        const {
-            firstPageRequest,
-            secondPageRequest,
-            firstPageResponse,
-            secondPageResponse,
-        } = new ListObjectVersionsFixtures()
+        const { firstPageRequest, secondPageRequest, firstPageResponse, secondPageResponse } =
+            new ListObjectVersionsFixtures()
 
         it('lists objects and their versions with a continuation token for the next page of results', async function () {
             when(mockS3.listObjectVersions(deepEqual(firstPageRequest))).thenReturn(success(firstPageResponse))
